+++ NEWS.md
@@ -3,6 +3,8 @@
 * Updated `decimateWavFiles` to work with wav files with bit rates different
 than 16 (bugs in `seewave` functions do not pass along bit rate properly)
 
+* Bug fix in `matchEnvData` for datasets that had no `UTC` axis
+
 # PAMmisc 1.8.0
 
 * Reworking of enviro functions to make HYCOM datasets easier to use. Will now