+++ NEWS.md
@@ -5,6 +5,9 @@ than 16 (bugs in `seewave` functions do not pass along bit rate properly)
 
 * Bug fix in `matchEnvData` for datasets that had no `UTC` axis
 
+* Changed defaults in `matchEnvData` and `ncToData` to only calculate `mean`, not
+`median` and `sd`
+
 # PAMmisc 1.8.0
 
 * Reworking of enviro functions to make HYCOM datasets easier to use. Will now